boots and runs on Linux now too

+++ hashstr.cxx
@@ -41,6 +41,8 @@
 #include <string.h>
 #include <sys/types.h>
 
+#include "mud.hxx"
+
 #if !defined(FALSE)
 #define FALSE 0
 #endif
@@ -191,7 +193,7 @@ void hash_dump(int hash)
     psize = sizeof(struct hashstr_data);
     for (c = 0, ptr = string_hash[hash]; ptr; ptr = ptr->next, c++)
     {
-        str = (char *)(((int)ptr) + psize);
+        str = ((char *)ptr) + psize;
         fprintf(stderr, "Len:%4d Lnks:%5d Str: %s\n\r", ptr->length, ptr->links, str);
     }
     fprintf(stderr, "Total strings in hash %d: %d\n\r", hash, c);
@@ -260,7 +262,7 @@ void show_high_hash(int top)
         for (ptr = string_hash[x]; ptr; ptr = ptr->next)
             if (ptr->links >= top)
             {
-                str = (char *)(((int)ptr) + psize);
+                str = ((char*)ptr) + psize;
                 fprintf(stderr, "Links: %5d  String: >%s<\n\r", ptr->links, str);
             }
 }
